asg2.2/requirements.txt
Flask==2.0.2
Flask_Cors==3.0.10
matplotlib==3.4.3
numpy==1.21.4
Pillow==8.4.0
six==1.16.0
tensorflow-cpu==2.7.0
tensorflow_hub==0.12.0
Werkzeug==2.0.2